separate components on files: MobileBreakpoint, DesktopBreakpoint and DynamicButtons
